+++ src/kdtree/index.rs
@@ -5,37 +5,37 @@ use crate::kdtree::constants::{KDBUSH_HEADER_SIZE, KDBUSH_MAGIC, KDBUSH_VERSION}
 use crate::kdtree::error::KdbushError;
 
 #[derive(Debug, Clone, PartialEq)]
-pub struct OwnedKdbush {
+pub struct OwnedKDTree {
     pub(crate) buffer: Vec<u8>,
     pub(crate) node_size: usize,
     pub(crate) num_items: usize,
 }
 
-impl OwnedKdbush {
+impl OwnedKDTree {
     pub fn into_inner(self) -> Vec<u8> {
         self.buffer
     }
 
-    pub fn as_kdbush(&self) -> KdbushRef {
-        KdbushRef::try_new(self).unwrap()
+    pub fn as_kdbush(&self) -> KDTreeRef {
+        KDTreeRef::try_new(self).unwrap()
     }
 }
 
-impl AsRef<[u8]> for OwnedKdbush {
+impl AsRef<[u8]> for OwnedKDTree {
     fn as_ref(&self) -> &[u8] {
         &self.buffer
     }
 }
 
 #[derive(Debug, Clone, PartialEq)]
-pub struct KdbushRef<'a> {
+pub struct KDTreeRef<'a> {
     pub(crate) coords: &'a [f64],
     pub(crate) ids: Indices<'a>,
     pub(crate) node_size: usize,
     pub(crate) num_items: usize,
 }
 
-impl<'a> KdbushRef<'a> {
+impl<'a> KDTreeRef<'a> {
     pub fn try_new<T: AsRef<[u8]>>(data: &'a T) -> Result<Self, KdbushError> {
         let data = data.as_ref();
