+++ src/kdtree/mod.rs
@@ -4,9 +4,9 @@ pub mod error;
 pub mod index;
 pub mod r#trait;
 
-pub use builder::KdbushBuilder;
-pub use index::{KdbushRef, OwnedKdbush};
-pub use r#trait::KdbushIndex;
+pub use builder::KDTreeBuilder;
+pub use index::{KDTreeRef, OwnedKDTree};
+pub use r#trait::KDTreeIndex;
 
 #[cfg(test)]
 mod test;
