+++ src/lib.rs
@@ -1,12 +1,12 @@
 #![doc = include_str!("../README.md")]
 
+mod error;
 pub mod indices;
 pub mod kdtree;
 pub mod rtree;
 pub mod r#type;
 
-pub use kdtree::{KdbushBuilder, KdbushIndex, KdbushRef, OwnedKdbush};
-pub use rtree::{OwnedRTree, RTreeBuilder, RTreeIndex, RTreeRef};
+pub use error::GeoIndexError;
 
 #[cfg(test)]
 pub(crate) mod test;
