+++ src/rtree/mod.rs
@@ -1,6 +1,5 @@
 pub mod builder;
 pub mod constants;
-pub mod error;
 pub mod index;
 pub mod sort;
 pub mod r#trait;
@@ -10,4 +9,4 @@ pub mod util;
 pub use builder::RTreeBuilder;
 pub use index::{OwnedRTree, RTreeRef};
 pub use r#trait::RTreeIndex;
-pub use sort::HilbertSort;
+pub use sort::{HilbertSort, STRSort};
