+++ src/test/integration.rs
@@ -3,7 +3,7 @@ use std::fs::read;
 use bytemuck::cast_slice;
 
 use crate::rtree::HilbertSort;
-use crate::{OwnedRTree, RTreeBuilder, RTreeRef};
+use crate::rtree::{OwnedRTree, RTreeBuilder, RTreeRef};
 
 fn create_flatbush_from_data_path(data_path: &str) -> OwnedRTree<f64> {
     let buffer = read(data_path).unwrap();
